requirements-dev.txt
-r requirements.txt
coverage==4.5.1
Faker==0.8.15
flake8==3.5.0
mccabe==0.6.1
py==1.5.3
pycodestyle==2.3.1
pyflakes==2.0.0
pytest==3.6.0
pytest-cov==2.5.1
pytest-django==3.2.1
freezegun==0.3.10
pytest-freezegun==0.2.0
model-mommy==1.6.0
text_unidecode==1.2
pytest-mock==1.10.0
